// src/modules/ui/components/FormCalendar/interfaces.ts
import { ReactNode } from 'react';
import { StyleProp, ViewStyle, TextStyle } from 'react-native';

/* Defining the props that the component will accept. */
export interface FormCalendarProps {
    editable?: boolean;
    icon?: ReactNode;
    inputDateFormat: string;
    label: string;
    labelStyle?: StyleProp<TextStyle>;
    minDate?: string;
    name: string;
    onChangeDate?: (date: string) => void;
    style?: StyleProp<ViewStyle>;
}